Instalar boostrap ycomenzar con la vista del mesero. Close #47

+++ src/components/Products/Products.jsx
@@ -1,10 +1,11 @@
 import React, { useState, useEffect } from "react";
 import { db } from "../../firebaseConfig/FirebaseConfig";
-import { collection, getDocs, getDoc, doc, deleteDoc } from "firebase/firestore";
+import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
 import './Products.css';
 import Navbar from "../Navbar/Navbar"
 import { GrTrash } from "react-icons/gr";
 import {useNavigate} from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 
@@ -12,7 +13,6 @@ const Products = () => {
     const [products, setProducts] = useState([])
     console.log('holaa1 estamos iniciando');
     const productsLunch = collection(db, 'Lunch')
-    const productsBreakfast = collection(db, 'Breakfast')
     console.log('holaaa2 estamos en lunch');
 
     const navigate = useNavigate();
@@ -28,9 +28,26 @@ const Products = () => {
     };
     
     const deleteProduct = async (id) => {
-        const productDoc = doc(db, 'Lunch', id)
-        await deleteDoc(productDoc)
-        getLunch()
+        Swal.fire({
+            title: 'Esta seguro que quiere eliminarlo',
+            text: "No habra forma de recuperar el producto eliminado",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, eliminalo!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                const productDoc = doc(db, 'Lunch', id)
+                deleteDoc(productDoc)
+                getLunch()
+              Swal.fire({
+                title:'Eliminado!',
+                text:'Producto Eliminado',
+                imageUrl: "https://cliply.co/wp-content/uploads/2021/03/372103860_CHECK_MARK_400px.gif",
+            })
+            }
+          })
     }
     useEffect(() => {
         getLunch()
@@ -43,10 +60,11 @@ const Products = () => {
             <button onClick={() => navigate("/Breakfast")}>Desayuno</button>            {
                 products.map(list => (
                     <div key={list.id} className='containerCard'>
-                        <p className='productName'>{list.ProductName}</p>
+                        <h5 className='productName'>{list.ProductName}</h5>
                         <p className='productPrice'>Precio: ${list.ProductPrice}</p>
                         <img className='productImg'src={list.ProductImg} alt='' />
-                        <button onClick={() => {deleteProduct(list.id)}} className='btnDelete'><GrTrash /></button>
+                        <br/>
+                        <button onClick={() => {deleteProduct(list.id)}} className="btn btn-danger"><GrTrash /></button>
                         <p></p>
                     </div>
                 ))
